+++ CHANGELOG.md
@@ -11,11 +11,13 @@
 - System refactor: Added subscriber runner
 - Fix: Class notifications can only be listened by class subscriptions
 - Refactor: Removed `publish_model_data` to have a unique model publisher `ps_publish`
-- Refactor: Renamed `ps_before_sync` into `ps_before_publish`, `ps_skip_sync` into `ps_skip_publish`, `ps_after_sync` into `ps_after_publish`
+- Refactor: Renamed `ps_before_sync` into `ps_before_publish`, `ps_after_sync` into `ps_after_publish`
 - Refactor: Renamed `payload.attributes` into `payload.info`
 - Feat: Support for plain Ruby Objects (Non ActiveRecord models)
 - Fix: Retry errors for 5 times before exiting notifications listener
-- Feat: Added transactions max_buffer 
+- Feat: Added transactions max_buffer
+- Feat: add `ps_perform_publish` to perform the callback for a specific action 
+- Feat: Removed `ps_skip_sync` callback
 
 # 0.6.0 (March 03, 2021)
 - feat: add support to include custom payload headers